[TAN-5512] Changed homepagebuilder to use contentBuilderLayout
