+++ front/app/api/content_builder/useContentBuilderLayout.test.ts
@@ -12,6 +12,7 @@ const projectApiPath =
   '*/projects/:projectId/content_builder_layouts/project_description';
 const folderApiPath =
   '*/project_folders/:folderId/content_builder_layouts/project_folder_description';
+const homepageApiPath = '*home_pages/content_builder_layouts/homepage';
 
 const server = setupServer(
   http.get(projectApiPath, () => {
@@ -47,6 +48,33 @@ describe('useContentBuilderLayout', () => {
   it('returns data correctly for a folder', async () => {
     server.use(
       http.get(folderApiPath, () => {
+        return HttpResponse.json(
+          { data: contentBuilderLayoutData },
+          { status: 200 }
+        );
+      })
+    );
+
+    const spy = jest.spyOn(global, 'fetch');
+    const { result } = renderHook(
+      () => useContentBuilderLayout('folder', 'folderId'),
+      {
+        wrapper: createQueryClientWrapper(),
+      }
+    );
+
+    expect(result.current.isLoading).toBe(true);
+
+    await waitFor(() => expect(result.current.isSuccess).toBe(true));
+
+    expect(result.current.isLoading).toBe(false);
+    expect(result.current.data?.data).toEqual(contentBuilderLayoutData);
+    expect(spy).toHaveBeenCalledTimes(1);
+  });
+
+  it('returns data correctly for the homepage', async () => {
+    server.use(
+      http.get(homepageApiPath, () => {
         return HttpResponse.json(
           { data: contentBuilderLayoutData },
           { status: 200 }
