--- front/app/api/home_page_layout/__mocks__/homepageLayout.ts
@@ -1,9 +1,0 @@
-export const homepageBuilderLayoutData = {
-  id: '058898ed-46b6-4527-ad8b-098604b4de2d',
-  type: 'content_builder_layout',
-  attributes: {
-    enabled: true,
-    craftjs_json: {},
-    code: 'test',
-  },
-};
