--- front/app/api/home_page_layout/keys.ts
@@ -1,10 +1,0 @@
-import { QueryKeys } from 'utils/cl-react-query/types';
-
-const baseKey = { type: 'content_builder_layout' };
-
-const homepageBuilderKeys = {
-  all: () => [baseKey],
-  items: () => [{ ...baseKey, operation: 'item' }],
-} satisfies QueryKeys;
-
-export default homepageBuilderKeys;
