--- front/app/api/home_page_layout/useHomepageLayout.test.ts
@@ -1,56 +1,0 @@
-import { http, HttpResponse } from 'msw';
-import { setupServer } from 'msw/node';
-
-import createQueryClientWrapper from 'utils/testUtils/queryClientWrapper';
-import { renderHook, waitFor } from 'utils/testUtils/rtl';
-
-import { homepageBuilderLayoutData } from './__mocks__/homepageLayout';
-import useHomepageLayout from './useHomepageLayout';
-
-const apiPath = '*home_pages/content_builder_layouts/homepage';
-
-const server = setupServer(
-  http.get(apiPath, () => {
-    return HttpResponse.json(
-      { data: homepageBuilderLayoutData },
-      { status: 200 }
-    );
-  })
-);
-
-describe('useHomepageLayout', () => {
-  beforeAll(() => server.listen());
-  afterAll(() => server.close());
-
-  it('returns data correctly', async () => {
-    const spy = jest.spyOn(global, 'fetch');
-    const { result } = renderHook(() => useHomepageLayout(), {
-      wrapper: createQueryClientWrapper(),
-    });
-
-    expect(result.current.isLoading).toBe(true);
-
-    await waitFor(() => expect(result.current.isSuccess).toBe(true));
-
-    expect(result.current.isLoading).toBe(false);
-    expect(result.current.data?.data).toEqual(homepageBuilderLayoutData);
-    expect(spy).toHaveBeenCalledTimes(1);
-  });
-
-  it('returns error correctly', async () => {
-    server.use(
-      http.get(apiPath, () => {
-        return HttpResponse.json(null, { status: 500 });
-      })
-    );
-
-    const { result } = renderHook(() => useHomepageLayout(), {
-      wrapper: createQueryClientWrapper(),
-    });
-
-    expect(result.current.isLoading).toBe(true);
-    await waitFor(() => expect(result.current.isError).toBe(true));
-    expect(result.current.error).toBeDefined();
-    expect(result.current.isLoading).toBe(false);
-  });
-});
