+++ front/app/api/project_folders/useDeleteProjectFolder.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import adminPublicationsKeys from 'api/admin_publications/keys';
 import adminPublicationsStatusCountsKeys from 'api/admin_publications_status_counts/keys';
-import homepageBuilderKeys from 'api/home_page_layout/keys';
+import contentBuilderKeys from 'api/content_builder/keys';
 import projectFoldersMiniKeys from 'api/project_folders_mini/keys';
 import projectsKeys from 'api/projects/keys';
 
@@ -42,7 +42,7 @@ const useDeleteProjectFolder = () => {
         queryKey: adminPublicationsStatusCountsKeys.items(),
       });
       queryClient.invalidateQueries({
-        queryKey: homepageBuilderKeys.all(),
+        queryKey: contentBuilderKeys.all(),
       });
     },
   });
