+++ front/app/api/projects/useDeleteProject.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import adminPublicationsKeys from 'api/admin_publications/keys';
 import adminPublicationsStatusCountsKeys from 'api/admin_publications_status_counts/keys';
-import homepageBuilderKeys from 'api/home_page_layout/keys';
+import contentBuilderKeys from 'api/content_builder/keys';
 import projectsMiniAdminKeys from 'api/projects_mini_admin/keys';
 
 import fetcher from 'utils/cl-react-query/fetcher';
@@ -30,7 +30,7 @@ const useDeleteProject = () => {
         queryKey: adminPublicationsStatusCountsKeys.items(),
       });
       queryClient.invalidateQueries({
-        queryKey: homepageBuilderKeys.all(),
+        queryKey: contentBuilderKeys.all(),
       });
       queryClient.invalidateQueries({
         queryKey: projectsMiniAdminKeys.lists(),
